[[ RefactorSyntax ]] Moved keywords exec methods to their own file.
