+++ engine/src/statemnt.cpp
@@ -533,215 +533,7 @@ Exec_stat MCComref::exec(MCExecPoint &ep)
 
 void MCComref::exec_ctxt(MCExecContext& ctxt)
 {
-    MCKeywordsExecCommand(ctxt, resolved, handler, params, name, line, pos, platform_message, false);
+    MCKeywordsExecCommandOrFunction(ctxt, resolved, handler, params, name, line, pos, platform_message, false);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-
-Exec_stat MCKeywordsExecuteStatements(MCExecContext& ctxt, MCStatement *p_statements, Exec_errors p_error, bool p_is_try)
-{
-    Exec_stat stat = ES_NORMAL;
-    MCStatement *tspr = p_statements;
-    while (tspr != NULL)
-    {
-        if (MCtrace || MCnbreakpoints)
-        {
-            MCB_trace(ctxt, tspr->getline(), tspr->getpos());
-            if (MCexitall)
-                break;
-        }
-        ctxt . GetEP() . setline(tspr->getline());
-        
-        stat = tspr->exec(ctxt . GetEP());
-        
-        // MW-2011-08-17: [[ Redraw ]] Flush any screen updates.
-        MCRedrawUpdateScreen();
-        
-        switch(stat)
-        {
-            case ES_NORMAL:
-                if (MCexitall)
-                    return ES_NORMAL;
-                tspr = tspr->getnext();
-                break;
-            case ES_NEXT_REPEAT:
-                tspr = NULL;
-                break;
-            case ES_EXIT_REPEAT:
-                return stat;
-            case ES_EXIT_SWITCH:
-                return ES_NORMAL;
-            case ES_ERROR:
-                if ((MCtrace || MCnbreakpoints) && !MCtrylock && !MClockerrors)
-                    do
-                    {
-                        ctxt . IgnoreLastError();
-                        MCB_error(ctxt, tspr->getline(), tspr->getpos(),
-                                  EE_REPEAT_BADSTATEMENT);
-                    }
-                while (MCtrace && (stat = tspr->exec(ctxt . GetEP())) != ES_NORMAL);
-                if (stat == ES_ERROR)
-                {
-                    if (MCexitall)  
-                        return p_is_try ? ES_NORMAL : ES_ERROR;
-                    
-                    ctxt . LegacyThrow(p_error);
-                    return stat;
-                }
-                else
-                    tspr = tspr->getnext();
-                break;
-            default:
-                return stat;
-        }
-    }
-    return stat;
-}
-
-void MCKeywordsExecCommand(MCExecContext& ctxt, bool resolved, MCHandler *handler, MCParameter *params, MCNameRef name, uint2 line, uint2 pos, bool platform_message, bool is_function)
-{
-	if (MCscreen->abortkey())
-	{
-		ctxt . LegacyThrow(EE_HANDLER_ABORT);
-		return;
-	}
-    
-	if (!resolved)
-	{
-		// MW-2008-01-28: [[ Inherited parentScripts ]]
-		// If we are in parentScript context, then the object we search for
-		// private handlers in is the parentScript's object, rather than the
-		// ep's.
-		MCParentScriptUse *t_parentscript;
-		t_parentscript = ctxt . GetParentScript();
-        
-		MCObject *t_object;
-		if (t_parentscript == NULL)
-			t_object = ctxt . GetObject();
-        else
-            t_object = t_parentscript -> GetParent() -> GetObject();
-        
-        // MW-2008-10-28: [[ ParentScripts ]] Private handlers are resolved
-        //   relative to the object containing the handler we are executing.
-        MCHandler *t_resolved_handler;
-		t_resolved_handler = t_object -> findhandler(is_function ? HT_FUNCTION : HT_MESSAGE, name);
-		if (t_resolved_handler != NULL && t_resolved_handler -> isprivate())
-			handler = t_resolved_handler;
-        
-        resolved = true;
-    }
-
-    if (is_function)
-        MCexitall = False;
-    
-	// Go through all the parameters to the function, if they are not variables, clear their current value. Each parameter stores an expression
-	// which allows its value to be re-evaluated in a given context. Re-evaluate each in the context of ep and set it to the new value.
-	// As the ep should contain the context of the caller at this point, the expression should be evaluated in that context.
-    Exec_stat stat;
-	MCParameter *tptr = params;
-	while (tptr != NULL)
-	{
-		MCVariable* t_var;
-		t_var = tptr -> evalvar(ctxt);
-        
-		if (t_var == NULL)
-		{
-			tptr -> clear_argument();
-            MCAutoValueRef t_value;
-			if (!ctxt . TryToEvaluateParameter(tptr, line, pos, is_function ? EE_FUNCTION_BADSOURCE : EE_STATEMENT_BADPARAM, &t_value))
-                return;
-			tptr->setvalueref_argument(*t_value);
-		}
-		else
-			tptr->set_argument_var(t_var);
-        
-        tptr = tptr->getnext();
-    }
-	MCObject *p = ctxt . GetObject();
-	MCExecContext *oldctxt = MCECptr;
-	MCECptr = &ctxt;
-	stat = ES_NOT_HANDLED;
-	Boolean added = False;
-	if (MCnexecutioncontexts < MAX_CONTEXTS)
-	{
-		ctxt . GetEP() . setline(line);
-		MCexecutioncontexts[MCnexecutioncontexts++] = &ctxt;
-		added = True;
-	}
-    
-#ifdef _MOBILE
-    if (platform_message)
-    {
-        stat = MCHandlePlatformMessage(name, params);
-    }
-#endif
-    
-	if (handler != nil)
-	{
-        // MW-2008-10-28: [[ ParentScripts ]] If we are in the context of a
-        //   parent, then use a special method.
-        if (ctxt . GetParentScript() == nil)
-            stat = p -> exechandler(handler, params);
-        else
-            stat = p -> execparenthandler(handler, params, ctxt . GetParentScript());
-        
-        switch(stat)
-        {
-            case ES_ERROR:
-            case ES_PASS:
-                MCeerror->add(is_function ? EE_FUNCTION_BADFUNCTION : EE_STATEMENT_BADCOMMAND, line, pos, handler -> getname());
-                if (MCerrorptr == NULL)
-                    MCerrorptr = p;
-                stat = ES_ERROR;
-                break;
-                
-            case ES_EXIT_HANDLER:
-                stat = ES_NORMAL;
-                break;
-                
-            default:
-                break;
-        }
-	}
-	else
-	{
-		stat = MCU_dofrontscripts(is_function ? HT_FUNCTION : HT_MESSAGE, name, params);
-		Boolean olddynamic = MCdynamicpath;
-		MCdynamicpath = MCdynamiccard != NULL;
-		if (stat == ES_PASS || stat == ES_NOT_HANDLED)
-        {
-            if (is_function)
-            {
-                // PASS STATE FIX
-                Exec_stat oldstat = stat;
-                stat = p->handle(HT_FUNCTION, name, params, p);
-                if (oldstat == ES_PASS && stat == ES_NOT_HANDLED)
-                    stat = ES_PASS;
-            }
-            else
-            {
-                switch (stat = p->handle(HT_MESSAGE, name, params, p))
-                {
-                    case ES_ERROR:
-                    case ES_NOT_FOUND:
-                    case ES_NOT_HANDLED:
-                    case ES_PASS:
-                        MCeerror->add(EE_STATEMENT_BADCOMMAND, line, pos, name);
-                        stat = ES_ERROR;
-                        break;
-                    case ES_EXIT_HANDLER:
-                        stat = ES_NORMAL;
-                        break;
-                    default:
-                        break;
-                }
-            }
-        }
-            
-		MCdynamicpath = olddynamic;
-	}
-	MCECptr = oldctxt;
-	if (added)
-		MCnexecutioncontexts--;
-	ctxt . SetExecStat(stat);
-}
