+++ engine/src/statemnt.h
@@ -85,7 +85,4 @@ public:
 	virtual void exec_ctxt(MCExecContext&);
 };
 
-Exec_stat MCKeywordsExecuteStatements(MCExecContext& ctxt, MCStatement *p_statements, Exec_errors p_error, bool p_is_try = false);
-void MCKeywordsExecCommand(MCExecContext& ctxt, bool resolved, MCHandler *handler, MCParameter *params, MCNameRef name, uint2 line, uint2 pos, bool platform_message, bool is_function);
-
 #endif
